requirements.txt
asgiref>=3.8.1
black==24.2.0
boto3==1.34.69
celery==5.3.6
comtypes==1.3.0
coverage==7.4.1
Django==5.1.6
django-axes==6.3.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.15.0
djangorestframework-simplejwt
drf-yasg==1.21.7
environ
flake8==7.0.0
gunicorn==23.0.0
isort==5.13.2
Markdown==3.5.1
pillow==10.2.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
pywifi==1.1.12
pytest==8.0.1
pytest-django==4.8.0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.0.1
requests
sentry-sdk==1.40.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
